# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: AI 工具市场 V1 API
  version: 0.3.0
  description: |
    AI 工具市场 V1 接口契约，用于用户端、管理端、Worker 和自动化测试联调?

    本文件按 `项目规划/V1-7天交付给组员弢发包` 中的 V1 范围整理?
    - 覆盖健康棢查登录鉴权工具市场任务算力管理端?Worker 内部回写主链路?
    - V1 不声明订单套餐支付审批多模开?API 等未纳入本周验收的能力?
    - 自动化测试可读取根节?`x-test-scenarios` 执行 P0 场景?
  contact:
    name: AI miniLab
servers:
  - url: http://localhost:8080
    description: Docker Compose 本地后端
...
paths:
...
  /api/admin/v1/agent/model-config:
    get:
      tags: [AdminModelConfig]
      summary: Get default model API config
      operationId: getAgentModelConfig
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentModelConfigApiResponse"
    put:
      tags: [AdminModelConfig]
      summary: Save default model API config
      operationId: saveAgentModelConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentModelConfigRequest"
      responses:
        "200":
          description: Saved model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentModelConfigApiResponse"
    post:
      tags: [AdminModelConfig]
      summary: Create model API config
      operationId: createAgentModelConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentModelConfigRequest"
      responses:
        "200":
          description: Created model config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentModelConfigApiResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiToken:
      type: apiKey
      in: header
      name: X-Internal-Signature
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - SUCCESS
        - PARAM_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - ADMIN_UNAUTHORIZED
        - ADMIN_FORBIDDEN
        - TOOL_NOT_FOUND
        - TOOL_OFFLINE
        - CREDIT_NOT_ENOUGH
        - TASK_NOT_FOUND
        - TASK_STATUS_INVALID
        - MODEL_CALL_FAILED
        - SYSTEM_ERROR
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          example: ok
        traceId:
          type: string
          nullable: true
          description: 当前实现使用 traceId；旧规划中的 requestId 视为同类追踪字段?
...
    AgentModelConfigApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AgentModelConfig"
...
    AgentModelConfig:
      type: object
      properties:
        id:
          type: integer
          format: int64
        displayName:
          type: string
          nullable: true
        configCode:
          type: string
          nullable: true
        provider:
          type: string
        modelName:
          type: string
        baseUrl:
          type: string
          nullable: true
        apiKeyMasked:
          type: string
          nullable: true
        minimaxGroupId:
          type: string
          nullable: true
        consoleUrl:
          type: string
          nullable: true
        balanceUrl:
          type: string
          nullable: true
        docsUrl:
          type: string
          nullable: true
        timeoutSeconds:
          type: integer
        inputTokenPricePer1m:
          type: number
          format: double
        outputTokenPricePer1m:
          type: number
          format: double
        billingUnit:
          type: string
          enum: [TOKEN_PER_M, PER_CALL]
        unitPrice:
          type: number
          format: double
        enabled:
          type: boolean
        isDefault:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
          description: Execution capabilities supported by this credential set.
    AgentModelConfigRequest:
      type: object
      required: [provider, modelName]
      properties:
        displayName:
          type: string
        configCode:
          type: string
        provider:
          type: string
        modelName:
          type: string
        baseUrl:
          type: string
        apiKey:
          type: string
        minimaxGroupId:
          type: string
        consoleUrl:
          type: string
        balanceUrl:
          type: string
        docsUrl:
          type: string
        timeoutSeconds:
          type: integer
          minimum: 1
          maximum: 300
        inputTokenPricePer1m:
          type: number
          format: double
          description: Input token cost per 1M tokens.
        outputTokenPricePer1m:
          type: number
          format: double
          description: Output token cost per 1M tokens.
        billingUnit:
          type: string
          enum: [TOKEN_PER_M, PER_CALL]
        unitPrice:
          type: number
          format: double
          description: Cost per billable generated item when billingUnit is PER_CALL.
        enabled:
          type: boolean
        isDefault:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
...